backend/requirements.txt
Flask>=2.0
Flask-Admin>=1.6
Flask-RBAC>=0.1.5
SQLAlchemy>=1.4
Flask-SQLAlchemy>=2.5
Flask-Migrate>=3.1
python-dotenv>=0.19
Flask-Cors>=4.0.0
google-generativeai>=0.7.0
pytest>=6.2
